transformation https://ui5infra.only.sap/job/legacy-free-transformer-test-pipelines/121/

+++ src/sap.ui.table/src/sap/ui/table/Table.js
@@ -3794,6 +3794,10 @@ sap.ui.define([
 	Table.prototype._initSelectionPlugin = function() {
 		var oSelectionPlugin = this.getPlugin("sap.ui.table.plugins.SelectionPlugin");
 
+		if (this.isDestroyed() || this.isDestroyStarted()) {
+			return;
+		}
+
 		if (oSelectionPlugin) {
 			this._destroyLegacySelectionPlugin();
